fix(CTX-101.1): fix Windows compile errors and a Linux test race caught by CI
windows-latest failed to compile: the `windows` crate feature list was
missing Win32_Security/Win32_System_Threading (CreateJobObjectW and
JOBOBJECT_EXTENDED_LIMIT_INFORMATION were gated behind them), the
HANDLE(child.as_raw_handle() as isize) cast was wrong (as_raw_handle()
already returns *mut c_void), and JobHandle needed a manual Send/Sync impl
since Tauri's app.manage() requires it and HANDLE wraps a raw pointer.

ubuntu-latest compiled but TEST-002 failed: the intermediate process could
exit before the grandchild's prctl(PR_SET_PDEATHSIG) call had actually
run, since fork() returning doesn't mean the child has executed anything
yet. Fixed with a second pipe so the grandchild signals readiness before
the intermediate reports its pid and exits.

+++ core/tauri-rust/src/daemon.rs
@@ -49,10 +49,7 @@ pub fn spawn_daemon(app: &AppHandle, script_path: PathBuf) -> std::io::Result<Da
         use std::os::windows::io::AsRawHandle;
         use windows::Win32::Foundation::HANDLE;
 
-        crate::supervisor::windows::assign_process(
-            &job_handle,
-            HANDLE(child.as_raw_handle() as isize),
-        )?;
+        crate::supervisor::windows::assign_process(&job_handle, HANDLE(child.as_raw_handle()))?;
     }
 
     let stdin = child
